feat: shared 레벨 공용 UI 컴포넌트 분리 및 적용
