+++ src/shared/ui/Dialog/index.ts
@@ -1,0 +1,1 @@
+export { Dialog, DialogTrigger, DialogContent, DialogHeader, DialogTitle } from "./Dialog"
